bank-support-env/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bank-support-env"
version = "1.0.0"
description = "Multi-turn banking customer support RL environment for OpenEnv"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Hackathon Team"},
]
keywords = ["openenv", "rl", "banking", "customer-support", "reinforcement-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openenv-core",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pydantic>=2.0",
    "openai>=1.0.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "websockets>=11.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.24.0",
]

[tool.setuptools.packages.find]
include = ["bank_support_env*", "server*"]
